refactor(core): split graph/core.rs, correlation/data_flow/mod.rs, graph/procedures.rs into submodules
Pure code-move refactor (wave 2) — zero logic changes, facades preserve
all paths; parent private use bindings kept so child use super::* keeps
resolving (downward privacy inheritance).

- graph/core.rs (1641) -> core/{mod,graph,node,edge,ids,property_store,
  stats} — 19/19 tests preserved
- correlation/data_flow/mod.rs (1632 -> 39) -> +tracker,types,analyzer,
  optimization,statistics — 33/33 tests preserved
- graph/procedures.rs (1594) -> procedures/{mod,types,shortest_path,
  centrality,community,similarity,topology,custom,registry} — 12/12

Part of phase5_split-oversized-files-wave2 (rulebook task).
