+++ crates/nexus-core/src/graph/core/graph.rs
@@ -1,339 +1,29 @@
-//! Core graph data structures - Graph, Node, Edge
-//!
-//! This module provides high-level graph data structures that wrap the low-level
-//! storage records and provide a more user-friendly API for graph operations.
+//! Graph data structure and operations
 
+use super::edge::Edge;
+use super::ids::{EdgeId, NodeId};
+use super::node::Node;
+use super::property_store::PropertyStore;
+use super::stats::{DegreeStats, EdgeTypeStats, GraphStats, NodeTypeStats, PathStats};
 use crate::graph::simple::PropertyValue;
 use crate::storage::{NodeRecord, RecordStore, RelationshipRecord};
 use crate::{Error, Result};
 use parking_lot::RwLock;
-use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::sync::Arc;
-
-/// Property record for storage
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct PropertyRecord {
-    /// Key of the property
-    key: String,
-    /// Value of the property
-    value: PropertyValue,
-    /// Pointer to next property in chain (u64::MAX if last)
-    next_ptr: u64,
-}
-
-/// Property storage for managing property chains
-#[derive(Debug)]
-struct PropertyStore {
-    /// In-memory property storage (in real implementation, this would be persistent)
-    properties: HashMap<u64, PropertyRecord>,
-    /// Next available property pointer
-    next_ptr: u64,
-}
-
-impl PropertyStore {
-    fn new() -> Self {
-        Self {
-            properties: HashMap::new(),
-            next_ptr: 1,
-        }
-    }
-
-    /// Store a property chain and return the head pointer
-    fn store_properties(&mut self, properties: HashMap<String, PropertyValue>) -> u64 {
-        if properties.is_empty() {
-            return u64::MAX;
-        }
-
-        let mut current_ptr = u64::MAX;
-
-        // Store properties in reverse order to maintain chain structure
-        let mut prop_vec: Vec<_> = properties.into_iter().collect();
-        prop_vec.reverse();
-
-        for (key, value) in prop_vec {
-            let ptr = self.next_ptr;
-            self.next_ptr += 1;
-
-            let record = PropertyRecord {
-                key,
-                value,
-                next_ptr: current_ptr,
-            };
-
-            self.properties.insert(ptr, record);
-            current_ptr = ptr;
-        }
-
-        current_ptr
-    }
-
-    /// Load properties from a property chain
-    fn load_properties(&self, head_ptr: u64) -> Result<HashMap<String, PropertyValue>> {
-        let mut properties = HashMap::new();
-        let mut current_ptr = head_ptr;
-
-        while current_ptr != u64::MAX {
-            if let Some(record) = self.properties.get(&current_ptr) {
-                properties.insert(record.key.clone(), record.value.clone());
-                current_ptr = record.next_ptr;
-            } else {
-                return Err(Error::Storage(format!(
-                    "Property record not found at pointer {}",
-                    current_ptr
-                )));
-            }
-        }
-
-        Ok(properties)
-    }
-}
-
-/// A unique identifier for nodes in the graph
-#[derive(
-    Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord, serde::Serialize, serde::Deserialize,
-)]
-pub struct NodeId(pub u64);
-
-impl NodeId {
-    /// Create a new node ID
-    pub fn new(id: u64) -> Self {
-        Self(id)
-    }
-
-    /// Get the raw ID value
-    pub fn value(&self) -> u64 {
-        self.0
-    }
-}
-
-impl From<u64> for NodeId {
-    fn from(id: u64) -> Self {
-        Self::new(id)
-    }
-}
-
-impl From<NodeId> for u64 {
-    fn from(node_id: NodeId) -> Self {
-        node_id.0
-    }
-}
-
-/// A unique identifier for relationships in the graph
-#[derive(
-    Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord, serde::Serialize, serde::Deserialize,
-)]
-pub struct EdgeId(pub u64);
-
-impl EdgeId {
-    /// Create a new edge ID
-    pub fn new(id: u64) -> Self {
-        Self(id)
-    }
-
-    /// Get the raw ID value
-    pub fn value(&self) -> u64 {
-        self.0
-    }
-}
-
-impl From<u64> for EdgeId {
-    fn from(id: u64) -> Self {
-        Self::new(id)
-    }
-}
-
-impl From<EdgeId> for u64 {
-    fn from(edge_id: EdgeId) -> Self {
-        edge_id.0
-    }
-}
-
-/// A node in the graph with labels and properties
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct Node {
-    /// Unique identifier for this node
-    pub id: NodeId,
-    /// Labels associated with this node
-    pub labels: Vec<String>,
-    /// Properties of this node
-    pub properties: HashMap<String, PropertyValue>,
-}
-
-impl Node {
-    /// Create a new node with the given ID and labels
-    pub fn new(id: NodeId, labels: Vec<String>) -> Self {
-        Self {
-            id,
-            labels,
-            properties: HashMap::new(),
-        }
-    }
-
-    /// Create a new node with ID, labels, and properties
-    pub fn with_properties(
-        id: NodeId,
-        labels: Vec<String>,
-        properties: HashMap<String, PropertyValue>,
-    ) -> Self {
-        Self {
-            id,
-            labels,
-            properties,
-        }
-    }
-
-    /// Add a label to this node
-    pub fn add_label(&mut self, label: String) {
-        if !self.labels.contains(&label) {
-            self.labels.push(label);
-        }
-    }
-
-    /// Remove a label from this node
-    pub fn remove_label(&mut self, label: &str) -> bool {
-        if let Some(pos) = self.labels.iter().position(|l| l == label) {
-            self.labels.remove(pos);
-            true
-        } else {
-            false
-        }
-    }
-
-    /// Check if this node has a specific label
-    pub fn has_label(&self, label: &str) -> bool {
-        self.labels.contains(&label.to_string())
-    }
-
-    /// Set a property on this node
-    pub fn set_property(&mut self, key: String, value: PropertyValue) {
-        self.properties.insert(key, value);
-    }
-
-    /// Get a property from this node
-    pub fn get_property(&self, key: &str) -> Option<&PropertyValue> {
-        self.properties.get(key)
-    }
-
-    /// Remove a property from this node
-    pub fn remove_property(&mut self, key: &str) -> Option<PropertyValue> {
-        self.properties.remove(key)
-    }
-
-    /// Check if this node has a specific property
-    pub fn has_property(&self, key: &str) -> bool {
-        self.properties.contains_key(key)
-    }
-
-    /// Get all property keys
-    pub fn property_keys(&self) -> Vec<&String> {
-        self.properties.keys().collect()
-    }
-
-    /// Check if this node is empty (no labels and no properties)
-    pub fn is_empty(&self) -> bool {
-        self.labels.is_empty() && self.properties.is_empty()
-    }
-}
-
-/// An edge (relationship) in the graph connecting two nodes
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct Edge {
-    /// Unique identifier for this edge
-    pub id: EdgeId,
-    /// Source node ID
-    pub source: NodeId,
-    /// Target node ID
-    pub target: NodeId,
-    /// Type of this relationship
-    pub relationship_type: String,
-    /// Properties of this edge
-    pub properties: HashMap<String, PropertyValue>,
-}
-
-impl Edge {
-    /// Create a new edge with the given ID, source, target, and type
-    pub fn new(id: EdgeId, source: NodeId, target: NodeId, relationship_type: String) -> Self {
-        Self {
-            id,
-            source,
-            target,
-            relationship_type,
-            properties: HashMap::new(),
-        }
-    }
-
-    /// Create a new edge with ID, source, target, type, and properties
-    pub fn with_properties(
-        id: EdgeId,
-        source: NodeId,
-        target: NodeId,
-        relationship_type: String,
-        properties: HashMap<String, PropertyValue>,
-    ) -> Self {
-        Self {
-            id,
-            source,
-            target,
-            relationship_type,
-            properties,
-        }
-    }
-
-    /// Set a property on this edge
-    pub fn set_property(&mut self, key: String, value: PropertyValue) {
-        self.properties.insert(key, value);
-    }
-
-    /// Get a property from this edge
-    pub fn get_property(&self, key: &str) -> Option<&PropertyValue> {
-        self.properties.get(key)
-    }
-
-    /// Remove a property from this edge
-    pub fn remove_property(&mut self, key: &str) -> Option<PropertyValue> {
-        self.properties.remove(key)
-    }
-
-    /// Check if this edge has a specific property
-    pub fn has_property(&self, key: &str) -> bool {
-        self.properties.contains_key(key)
-    }
-
-    /// Get all property keys
-    pub fn property_keys(&self) -> Vec<&String> {
-        self.properties.keys().collect()
-    }
-
-    /// Check if this edge is empty (no properties)
-    pub fn is_empty(&self) -> bool {
-        self.properties.is_empty()
-    }
-
-    /// Get the other end of this edge given one node
-    pub fn other_end(&self, node_id: NodeId) -> Option<NodeId> {
-        if self.source == node_id {
-            Some(self.target)
-        } else if self.target == node_id {
-            Some(self.source)
-        } else {
-            None
-        }
-    }
-}
 
 /// A graph containing nodes and edges
 pub struct Graph {
     /// Storage backend for persistence
-    store: std::cell::RefCell<RecordStore>,
+    pub(super) store: std::cell::RefCell<RecordStore>,
     /// Catalog for label/type/key mappings
-    catalog: Arc<crate::catalog::Catalog>,
+    pub(super) catalog: Arc<crate::catalog::Catalog>,
     /// Cache of loaded nodes (in-memory)
-    node_cache: Arc<RwLock<HashMap<NodeId, Node>>>,
+    pub(super) node_cache: Arc<RwLock<HashMap<NodeId, Node>>>,
     /// Cache of loaded edges (in-memory)
-    edge_cache: Arc<RwLock<HashMap<EdgeId, Edge>>>,
+    pub(super) edge_cache: Arc<RwLock<HashMap<EdgeId, Edge>>>,
     /// Property storage for managing property chains
-    property_store: Arc<RwLock<PropertyStore>>,
+    pub(super) property_store: Arc<RwLock<PropertyStore>>,
 }
 
 impl Graph {
@@ -1176,466 +866,5 @@ impl std::fmt::Debug for Graph {
             .field("node_cache", &self.node_cache.read().len())
             .field("edge_cache", &self.edge_cache.read().len())
             .finish()
-    }
-}
-
-/// Helper structures for statistics calculation
-#[derive(Debug, Clone)]
-struct DegreeStats {
-    avg_degree: f64,
-    max_degree: usize,
-    min_degree: usize,
-}
-
-#[derive(Debug, Clone)]
-struct PathStats {
-    avg_shortest_path_length: f64,
-    diameter: usize,
-}
-
-#[derive(Debug, Clone)]
-struct NodeTypeStats {
-    isolated_nodes: usize,
-    leaf_nodes: usize,
-}
-
-#[derive(Debug, Clone)]
-struct EdgeTypeStats {
-    self_loops: usize,
-    bidirectional_edges: usize,
-}
-
-/// Statistics about the graph
-#[derive(Debug, Clone)]
-pub struct GraphStats {
-    /// Number of active nodes in the graph
-    pub total_nodes: usize,
-    /// Number of active edges in the graph
-    pub total_edges: usize,
-    /// Number of nodes in storage (including deleted)
-    pub storage_nodes: usize,
-    /// Number of edges in storage (including deleted)
-    pub storage_edges: usize,
-    /// Number of nodes in cache
-    pub cached_nodes: usize,
-    /// Number of edges in cache
-    pub cached_edges: usize,
-    /// Average degree (connections per node)
-    pub avg_degree: f64,
-    /// Maximum degree of any node
-    pub max_degree: usize,
-    /// Minimum degree of any node
-    pub min_degree: usize,
-    /// Graph density (actual edges / possible edges)
-    pub graph_density: f64,
-    /// Number of connected components
-    pub connected_components: usize,
-    /// Average clustering coefficient
-    pub avg_clustering_coefficient: f64,
-    /// Average shortest path length
-    pub avg_shortest_path_length: f64,
-    /// Graph diameter (longest shortest path)
-    pub diameter: usize,
-    /// Number of isolated nodes (degree 0)
-    pub isolated_nodes: usize,
-    /// Number of leaf nodes (degree 1)
-    pub leaf_nodes: usize,
-    /// Number of self-loops
-    pub self_loops: usize,
-    /// Number of bidirectional edges
-    pub bidirectional_edges: usize,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::testing::{create_isolated_test_graph, create_test_graph};
-
-    #[test]
-    fn test_node_creation() {
-        let (graph, _dir) = create_test_graph();
-
-        let node_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        assert_eq!(node_id.value(), 0);
-
-        let node = graph.get_node(node_id).unwrap().unwrap();
-        assert_eq!(node.id, node_id);
-        assert!(node.has_label("Person"));
-        assert_eq!(node.labels.len(), 1);
-    }
-
-    #[test]
-    fn test_node_with_multiple_labels() {
-        let (graph, _dir) = create_test_graph();
-
-        let node_id = graph
-            .create_node(vec!["Person".to_string(), "Employee".to_string()])
-            .unwrap();
-
-        let node = graph.get_node(node_id).unwrap().unwrap();
-        assert!(node.has_label("Person"));
-        assert!(node.has_label("Employee"));
-        assert_eq!(node.labels.len(), 2);
-    }
-
-    #[test]
-    fn test_node_properties() {
-        let (graph, _dir) = create_test_graph();
-
-        let node_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let mut node = graph.get_node(node_id).unwrap().unwrap();
-
-        node.set_property(
-            "name".to_string(),
-            PropertyValue::String("test".to_string()),
-        );
-        node.set_property("age".to_string(), PropertyValue::Int64(30));
-
-        assert!(node.has_property("name"));
-        assert!(node.has_property("age"));
-        assert_eq!(node.property_keys().len(), 2);
-    }
-
-    #[test]
-    fn test_edge_creation() {
-        let (graph, _dir) = create_test_graph();
-
-        let source_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let target_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-
-        let edge_id = graph
-            .create_edge(source_id, target_id, "KNOWS".to_string())
-            .unwrap();
-        assert_eq!(edge_id.value(), 0);
-
-        let edge = graph.get_edge(edge_id).unwrap().unwrap();
-        assert_eq!(edge.id, edge_id);
-        assert_eq!(edge.source, source_id);
-        assert_eq!(edge.target, target_id);
-        assert_eq!(edge.relationship_type, "KNOWS");
-    }
-
-    #[test]
-    fn test_edge_properties() {
-        let (graph, _dir) = create_test_graph();
-
-        let source_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let target_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let edge_id = graph
-            .create_edge(source_id, target_id, "KNOWS".to_string())
-            .unwrap();
-
-        let mut edge = graph.get_edge(edge_id).unwrap().unwrap();
-        edge.set_property("since".to_string(), PropertyValue::Int64(2020));
-
-        assert!(edge.has_property("since"));
-        assert_eq!(edge.property_keys().len(), 1);
-    }
-
-    #[test]
-    fn test_node_deletion() {
-        let (graph, _dir) = create_test_graph();
-
-        let node_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        assert!(graph.get_node(node_id).unwrap().is_some());
-
-        let deleted = graph.delete_node(node_id).unwrap();
-        assert!(deleted);
-        assert!(graph.get_node(node_id).unwrap().is_none());
-    }
-
-    #[test]
-    fn test_edge_deletion() {
-        let (graph, _dir) = create_test_graph();
-
-        let source_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let target_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let edge_id = graph
-            .create_edge(source_id, target_id, "KNOWS".to_string())
-            .unwrap();
-
-        assert!(graph.get_edge(edge_id).unwrap().is_some());
-
-        let deleted = graph.delete_edge(edge_id).unwrap();
-        assert!(deleted);
-        assert!(graph.get_edge(edge_id).unwrap().is_none());
-    }
-
-    #[test]
-    fn test_get_nodes_by_label() {
-        let (graph, _dir) = create_test_graph();
-
-        let _person1 = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let _person2 = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let _company = graph.create_node(vec!["Company".to_string()]).unwrap();
-
-        let person_nodes = graph.get_nodes_by_label("Person").unwrap();
-        assert_eq!(person_nodes.len(), 2);
-
-        let company_nodes = graph.get_nodes_by_label("Company").unwrap();
-        assert_eq!(company_nodes.len(), 1);
-    }
-
-    #[test]
-    fn test_get_edges_by_type() {
-        let (graph, _dir) = create_test_graph();
-
-        let person1 = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let person2 = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let company = graph.create_node(vec!["Company".to_string()]).unwrap();
-
-        let _knows_edge = graph
-            .create_edge(person1, person2, "KNOWS".to_string())
-            .unwrap();
-        let _works_edge = graph
-            .create_edge(person1, company, "WORKS_AT".to_string())
-            .unwrap();
-
-        let knows_edges = graph.get_edges_by_type("KNOWS").unwrap();
-        assert_eq!(knows_edges.len(), 1);
-
-        let works_edges = graph.get_edges_by_type("WORKS_AT").unwrap();
-        assert_eq!(works_edges.len(), 1);
-    }
-
-    #[test]
-    fn test_get_edges_for_node() {
-        let (graph, _dir) = create_test_graph();
-
-        let person1 = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let person2 = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let person3 = graph.create_node(vec!["Person".to_string()]).unwrap();
-
-        let _edge1 = graph
-            .create_edge(person1, person2, "KNOWS".to_string())
-            .unwrap();
-        let _edge2 = graph
-            .create_edge(person1, person3, "KNOWS".to_string())
-            .unwrap();
-
-        let edges = graph.get_edges_for_node(person1).unwrap();
-        assert_eq!(edges.len(), 2);
-    }
-
-    #[test]
-    fn test_graph_stats() {
-        let (graph, _dir) = create_test_graph();
-
-        let _person1 = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let _person2 = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let _edge = graph
-            .create_edge(NodeId::new(0), NodeId::new(1), "KNOWS".to_string())
-            .unwrap();
-
-        let stats = graph.stats().unwrap();
-        assert_eq!(stats.total_nodes, 2);
-        assert_eq!(stats.total_edges, 1);
-    }
-
-    #[test]
-    fn test_node_label_operations() {
-        let (graph, _dir) = create_test_graph();
-
-        let node_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let mut node = graph.get_node(node_id).unwrap().unwrap();
-
-        // Add label
-        node.add_label("Employee".to_string());
-        assert!(node.has_label("Employee"));
-        assert_eq!(node.labels.len(), 2);
-
-        // Remove label
-        let removed = node.remove_label("Person");
-        assert!(removed);
-        assert!(!node.has_label("Person"));
-        assert!(node.has_label("Employee"));
-        assert_eq!(node.labels.len(), 1);
-
-        // Try to remove non-existent label
-        let not_removed = node.remove_label("NonExistent");
-        assert!(!not_removed);
-    }
-
-    #[test]
-    fn test_edge_other_end() {
-        let (graph, _dir) = create_test_graph();
-
-        let source_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let target_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let edge_id = graph
-            .create_edge(source_id, target_id, "KNOWS".to_string())
-            .unwrap();
-
-        let edge = graph.get_edge(edge_id).unwrap().unwrap();
-
-        assert_eq!(edge.other_end(source_id), Some(target_id));
-        assert_eq!(edge.other_end(target_id), Some(source_id));
-        assert_eq!(edge.other_end(NodeId::new(999)), None);
-    }
-
-    #[test]
-    fn test_node_property_operations() {
-        let (graph, _dir) = create_test_graph();
-
-        let node_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let mut node = graph.get_node(node_id).unwrap().unwrap();
-
-        // Set properties
-        node.set_property(
-            "name".to_string(),
-            PropertyValue::String("test".to_string()),
-        );
-        node.set_property("age".to_string(), PropertyValue::Int64(30));
-
-        // Check properties
-        assert!(node.has_property("name"));
-        assert!(node.has_property("age"));
-        assert_eq!(node.property_keys().len(), 2);
-
-        // Get property
-        let age = node.get_property("age").unwrap();
-        assert_eq!(age, &PropertyValue::Int64(30));
-
-        // Remove property
-        let removed = node.remove_property("age");
-        assert_eq!(removed, Some(PropertyValue::Int64(30)));
-        assert!(!node.has_property("age"));
-        assert!(node.has_property("name"));
-    }
-
-    #[test]
-    fn test_edge_property_operations() {
-        let (graph, _dir) = create_test_graph();
-
-        let source_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let target_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let edge_id = graph
-            .create_edge(source_id, target_id, "KNOWS".to_string())
-            .unwrap();
-
-        let mut edge = graph.get_edge(edge_id).unwrap().unwrap();
-
-        // Set properties
-        edge.set_property("since".to_string(), PropertyValue::Int64(2020));
-        edge.set_property("strength".to_string(), PropertyValue::Float64(0.8));
-
-        // Check properties
-        assert!(edge.has_property("since"));
-        assert!(edge.has_property("strength"));
-        assert_eq!(edge.property_keys().len(), 2);
-
-        // Get property
-        let since = edge.get_property("since").unwrap();
-        assert_eq!(since, &PropertyValue::Int64(2020));
-
-        // Remove property
-        let removed = edge.remove_property("strength");
-        assert_eq!(removed, Some(PropertyValue::Float64(0.8)));
-        assert!(!edge.has_property("strength"));
-        assert!(edge.has_property("since"));
-    }
-
-    #[test]
-    fn test_property_chain_traversal() {
-        let (graph, _dir) = create_test_graph();
-
-        // Create a node with properties
-        let node_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let mut node = graph.get_node(node_id).unwrap().unwrap();
-
-        // Set multiple properties
-        node.set_property(
-            "name".to_string(),
-            PropertyValue::String("Alice".to_string()),
-        );
-        node.set_property("age".to_string(), PropertyValue::Int64(30));
-        node.set_property("active".to_string(), PropertyValue::Bool(true));
-
-        // Update the node to store properties
-        graph.update_node(node).unwrap();
-
-        // Retrieve the node and verify properties are loaded from the chain
-        let retrieved_node = graph.get_node(node_id).unwrap().unwrap();
-
-        // The properties should be loaded from the property chain
-        assert!(retrieved_node.has_property("name"));
-        assert!(retrieved_node.has_property("age"));
-        assert!(retrieved_node.has_property("active"));
-
-        assert_eq!(
-            retrieved_node.get_property("name"),
-            Some(&PropertyValue::String("Alice".to_string()))
-        );
-        assert_eq!(
-            retrieved_node.get_property("age"),
-            Some(&PropertyValue::Int64(30))
-        );
-        assert_eq!(
-            retrieved_node.get_property("active"),
-            Some(&PropertyValue::Bool(true))
-        );
-    }
-
-    #[test]
-    fn test_node_is_empty() {
-        let (graph, _dir) = create_test_graph();
-
-        let node_id = graph.create_node(vec![]).unwrap();
-        let node = graph.get_node(node_id).unwrap().unwrap();
-        assert!(node.is_empty());
-
-        let mut node_with_label = graph.get_node(node_id).unwrap().unwrap();
-        node_with_label.add_label("Person".to_string());
-        assert!(!node_with_label.is_empty());
-    }
-
-    // Uses the shared-catalog `create_test_graph` helper instead of
-    // `create_isolated_test_graph` — opening a fresh LMDB env here
-    // used to push the process past Windows' TLS-slot ceiling and
-    // surface as `Database(Mdb(TlsFull))`. This test only inspects
-    // the `edge.is_empty()` predicate on its own fresh node +
-    // relationship ids; it does not depend on the catalog label id
-    // space being empty, so the shared catalog is safe.
-    #[test]
-    fn test_edge_is_empty() {
-        let (graph, _dir) = create_test_graph();
-
-        let source_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let target_id = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let edge_id = graph
-            .create_edge(source_id, target_id, "KNOWS".to_string())
-            .unwrap();
-
-        let edge = graph.get_edge(edge_id).unwrap().unwrap();
-        assert!(edge.is_empty());
-
-        let mut edge_with_props = graph.get_edge(edge_id).unwrap().unwrap();
-        edge_with_props.set_property("since".to_string(), PropertyValue::Int64(2020));
-        assert!(!edge_with_props.is_empty());
-    }
-
-    #[test]
-    fn test_clear_cache() {
-        // Shared-catalog helper — see `test_edge_is_empty` above for
-        // the TLS rationale. This test only checks cache state on
-        // its own newly-created nodes + edge.
-        let (graph, _dir) = create_test_graph();
-
-        let node_id1 = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let node_id2 = graph.create_node(vec!["Person".to_string()]).unwrap();
-        let _edge_id = graph
-            .create_edge(node_id1, node_id2, "KNOWS".to_string())
-            .unwrap();
-
-        // Verify cache has entries
-        assert!(!graph.node_cache.read().is_empty());
-        assert!(!graph.edge_cache.read().is_empty());
-
-        // Clear cache
-        graph.clear_cache();
-
-        // Verify cache is empty
-        assert!(graph.node_cache.read().is_empty());
-        assert!(graph.edge_cache.read().is_empty());
     }
 }
